SH-30: Fix problem with sidebar on mobile devices, fix problem with footer
