+++ angularclient/src/app/modules/bills/bills.component.html
@@ -1,13 +1,13 @@
-<div class="calendar">
+<div class="calendar d-flex position-relative w-100 justify-content-center">
   <div
     *ngIf="events$ | async as evs; else loading"
-    class="row align-items-center"
+    class="row align-items-center w-100 g-0"
   >
     <h4 class="col-12 col-sm-4 col-md-4 col-lg-2 text-center text-sm-start">
       {{ viewDate | calendarDate: view + "ViewTitle":locale | titlecase }}
     </h4>
     <div
-      class="col-8 col-sm-6 col-md-6 col-lg-8 text-center justify-content-center"
+      class="col-8 col-sm-6 col-md-6 col-lg-8 text-center justify-content-center g-0"
     >
       <div class="row justify-content-center align-items-center">
         <fa-icon
@@ -45,7 +45,7 @@
     </div>
 
     <app-add-bill
-      class="col-4 col-sm-2 col-md-2 col-lg-2 text-end"
+      class="col-4 col-sm-2 col-md-2 col-lg-2 justify-content-end float-end text-end"
     ></app-add-bill>
 
     <mwl-calendar-month-view
