+++ angularclient/src/app/modules/bills/bills.component.scss
@@ -6,4 +6,5 @@
 
 .calendar {
   position: relative;
+  display: flex;
 }
