+++ angularclient/src/app/shared/components/navbar/navbar.component.scss
@@ -1,47 +1,9 @@
 @import "/src/assets/_variables.scss";
 
-nav {
+.navbar {
   padding: 0 1rem;
   height: 56px;
 }
-
-// .navbar {
-//   display: flex;
-//   justify-content: space-between;
-//   margin: 0 auto;
-//   width: 100%;
-//   padding: 0.2rem 1rem 0.2rem 1rem;
-//   border-bottom: 1px solid lightgrey;
-
-//   &__left {
-//     display: flex;
-//     flex-direction: row;
-//     gap: 7px;
-//     justify-content: center;
-//     align-items: center;
-
-//     .icon-wrapper {
-//       position: relative;
-
-//       fa-icon {
-//         font-size: 30px;
-//         text-align: center;
-//         vertical-align: middle;
-//       }
-
-//       .badge {
-//         width: auto;
-//         height: auto;
-//         margin: 0;
-//         border-radius: 50%;
-//         position: absolute;
-//         top: 0px;
-//         right: -4px;
-//         padding: 5px;
-//       }
-//     }
-//   }
-// }
 
 .account {
   display: flex;
@@ -55,4 +17,49 @@ nav {
 .logo {
   width: 200px;
   height: 30px;
+}
+
+.shared-home-icon {
+  font-size: 25px;
+}
+
+.hamburger {
+  position: relative;
+  top: calc(50% - 2px);
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 32px;
+}
+
+.hamburger > span,
+.hamburger > span::before,
+.hamburger > span::after {
+  display: block;
+  position: absolute;
+  width: 100%;
+  height: 3px;
+  border-radius: 99px;
+  background-color: #fff;
+  transition-duration: 0.25s;
+}
+
+.hamburger > span::before {
+  content: "";
+  top: -8px;
+}
+.hamburger > span::after {
+  content: "";
+  top: 8px;
+}
+
+.navbar.is-active .hamburger > span {
+  transform: rotate(45deg);
+}
+.navbar.is-active .hamburger > span::before {
+  top: 0;
+  transform: rotate(0deg);
+}
+.navbar.is-active .hamburger > span::after {
+  top: 0;
+  transform: rotate(90deg);
 }
