+++ angularclient/src/app/shared/components/navbar/navbar.component.ts
@@ -1,6 +1,6 @@
 import { ScreenSizeHelper } from 'app/core/helpers/screen-size-helper';
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { faBell } from '@fortawesome/free-solid-svg-icons';
+import { faBell, faBars, faXmark } from '@fortawesome/free-solid-svg-icons';
 import { AuthenticationService } from 'app/modules/identity/services/authentication.service';
 import { NotificationService } from 'app/modules/notifications/services/notification.service';
 import { Observable } from 'rxjs';
@@ -13,7 +13,8 @@ import { Observable } from 'rxjs';
 export class NavbarComponent implements OnInit {
   notificationIcon = faBell;
   notificationsCount$!: Observable<number>;
-  userFullName: string = 'FN';
+  userFullName: string = '';
+  menuToggled: boolean = false;
 
   constructor(
     public screenSizeHelper: ScreenSizeHelper,
@@ -24,6 +25,42 @@ export class NavbarComponent implements OnInit {
   ngOnInit(): void {
     this.notificationsCount$ = this.notificationService.notificationsCount$;
     this.userFullName = this.buildUserFullName();
+
+    const navItems = document.querySelectorAll('.nav-item');
+    Array.from(navItems).map((x) =>
+      x.addEventListener('click', () => {
+        this.toggleMenu();
+      })
+    );
+  }
+
+  toggleMenu(): void {
+    const sidebar = document.querySelector('.sidebar');
+    sidebar?.classList.toggle('is-active');
+
+    const navbar = document.querySelector('.navbar');
+    navbar?.classList.toggle('is-active');
+
+    this.menuToggled = !this.menuToggled;
+
+    const media = window.matchMedia('(max-width: 768px)');
+    if (media.matches) {
+      this.blurContent();
+    }
+  }
+
+  logout() {
+    this.authenticationService.logout();
+  }
+
+  private blurContent(): void {
+    const content = document.querySelector('.content') as HTMLDivElement;
+
+    if (this.menuToggled) {
+      content.style.opacity = '0.1';
+    } else {
+      content.style.opacity = '1';
+    }
   }
 
   private buildUserFullName(): string {
