+++ angularclient/src/app/shared/components/sidebar/sidebar.component.html
@@ -1,9 +1,3 @@
-<div class="menu-toggle" (click)="toggleMenu()">
-  <div class="menu-toggle__hamburger">
-    <span></span>
-  </div>
-</div>
-
 <div class="sidenav">
   <ul class="nav sidenav__top">
     <li class="nav-item">
